go dep migration, writable check for neo4j connection/node.
I had to remove one test from cyber_test.go, this was due to the need of an actual neo4j cluster running instead of a single node. Writable check depends on the role of the connected node and it uses the `dbms.cluster.role` procedure to be installed/enabled.